Refactor data collection and analysis; remove unused email demo and PDF conversion files; update email report period to 3 months; enhance CORS middleware configuration.
